Исправления наставника
