--- src/const.js
@@ -1,131 +1,0 @@
-export const DESTINATIONS = [
-  'Chamonix',
-  'Geneva',
-  'Amsterdam',
-  'Stuttgart',
-];
-
-export const ALL_TYPES_OFFERS = [
-  {
-    'type': 'Taxi',
-    'offers': [
-      {
-        'id': 'business',
-        'title': 'Upgrade to a business class',
-        'price': 120
-      }, {
-        'id': 'radio',
-        'title': 'Choose the radio station',
-        'price': 60
-      }, {
-        'id': 'uber',
-        'title': 'Order Uber',
-        'price': 20
-      }
-    ]
-  }, {
-    'type': 'Bus',
-    'offers': [
-      {
-        'id': 'luggage',
-        'title': 'Add luggage',
-        'price': 20
-      },{
-        'id': 'meal',
-        'title': 'Add meal',
-        'price': 10
-      }, {
-        'id': 'seats',
-        'title': 'Choose seats',
-        'price': 5
-      }
-    ]
-  }, {
-    'type': 'Train',
-    'offers': [
-      {
-        'id': 'luggage',
-        'title': 'Add luggage',
-        'price': 20
-      },{
-        'id': 'meal',
-        'title': 'Add meal',
-        'price': 10
-      }, {
-        'id': 'seats',
-        'title': 'Choose seats',
-        'price': 5
-      }, {
-        'id': 'comfort',
-        'title': 'Switch to comfort class',
-        'price': 90
-      }
-    ]
-  }, {
-    'type': 'Ship',
-    'offers': [
-      {
-        'id': 'car',
-        'title': 'Add car place',
-        'price': 120
-      },{
-        'id': 'meal',
-        'title': 'Add meal',
-        'price': 10
-      }, {
-        'id': 'comfort',
-        'title': 'Switch to comfort class',
-        'price': 70
-      }
-    ]
-  }, {
-    'type': 'Drive',
-    'offers': [
-      {
-        'id': 'transponder',
-        'title': 'Toll road transponder',
-        'price': 65
-      },
-      {
-        'id': 'car',
-        'title': 'Rent a car',
-        'price': 200
-      },
-    ]
-  }, {
-    'type': 'Flight',
-    'offers': [
-      {
-        'id': 'luggage',
-        'title': 'Add luggage',
-        'price': 30
-      }, {
-        'id': 'comfort',
-        'title': 'Switch to comfort class',
-        'price': 100
-      }, {
-        'id': 'meal',
-        'title': 'Add meal',
-        'price': 15
-      }, {
-        'id': 'seats',
-        'title': 'Choose seats',
-        'price': 5
-      }, {
-        'id': 'train',
-        'title': 'Travel by train',
-        'price': 40
-      }
-    ]
-  }, {
-    'type': 'Restaurant',
-    'offers': [
-      {
-        'id': 'book',
-        'title': 'Book a table',
-        'price': 50
-      }
-    ]
-  }
-];
-
