+++ src/main.js
@@ -7,7 +7,7 @@ import { sitePointTemplate } from './view/site-point-view';
 import { renderTemplate, RenderPosition } from './render';
 import { generateEvent, relationNameDescription } from './mock/event';
 
-const POINTS__COUNT = 12;
+const POINTS_COUNT = 12;
 
 const controlsNavigation = document.querySelector('.trip-controls__navigation');
 const controlsFilters = document.querySelector('.trip-controls__filters');
@@ -21,13 +21,13 @@ renderTemplate(mainSort, siteCreatePointListTemplate(), RenderPosition.BEFOREEND
 
 const itemsList = mainSort.querySelector('.trip-events__list');
 
-let events = Array.from({length: POINTS__COUNT}, generateEvent);
+let events = Array.from({length: POINTS_COUNT}, generateEvent);
 // FUNCTION FROM STACKOVERFLOW.. MERGED TWO ARRAYS BY KEY
 const mergeArrays = (arr1, arr2) => (arr1.map((x) => Object.assign(x, arr2.find((y) => y.destination === x.destination))));
 events = mergeArrays(events, relationNameDescription());
 
 renderTemplate(itemsList, siteCreatePointTemplate(events[0]), RenderPosition.BEFOREEND);
 
-for (let i = 1; i < POINTS__COUNT; i++) {
+for (let i = 1; i < POINTS_COUNT; i++) {
   renderTemplate(itemsList, sitePointTemplate(events[i]), RenderPosition.BEFOREEND);
 }
