+++ src/mock/event.js
@@ -1,4 +1,134 @@
-import { ALL_TYPES_OFFERS, DESTINATIONS } from '../const';
+export const DESTINATIONS = [
+  'Chamonix',
+  'Geneva',
+  'Amsterdam',
+  'Stuttgart',
+];
+
+export const ALL_TYPES_OFFERS = [
+  {
+    'type': 'Taxi',
+    'offers': [
+      {
+        'id': 'business',
+        'title': 'Upgrade to a business class',
+        'price': 120
+      }, {
+        'id': 'radio',
+        'title': 'Choose the radio station',
+        'price': 60
+      }, {
+        'id': 'uber',
+        'title': 'Order Uber',
+        'price': 20
+      }
+    ]
+  }, {
+    'type': 'Bus',
+    'offers': [
+      {
+        'id': 'luggage',
+        'title': 'Add luggage',
+        'price': 20
+      },{
+        'id': 'meal',
+        'title': 'Add meal',
+        'price': 10
+      }, {
+        'id': 'seats',
+        'title': 'Choose seats',
+        'price': 5
+      }
+    ]
+  }, {
+    'type': 'Train',
+    'offers': [
+      {
+        'id': 'luggage',
+        'title': 'Add luggage',
+        'price': 20
+      },{
+        'id': 'meal',
+        'title': 'Add meal',
+        'price': 10
+      }, {
+        'id': 'seats',
+        'title': 'Choose seats',
+        'price': 5
+      }, {
+        'id': 'comfort',
+        'title': 'Switch to comfort class',
+        'price': 90
+      }
+    ]
+  }, {
+    'type': 'Ship',
+    'offers': [
+      {
+        'id': 'car',
+        'title': 'Add car place',
+        'price': 120
+      },{
+        'id': 'meal',
+        'title': 'Add meal',
+        'price': 10
+      }, {
+        'id': 'comfort',
+        'title': 'Switch to comfort class',
+        'price': 70
+      }
+    ]
+  }, {
+    'type': 'Drive',
+    'offers': [
+      {
+        'id': 'transponder',
+        'title': 'Toll road transponder',
+        'price': 65
+      },
+      {
+        'id': 'car',
+        'title': 'Rent a car',
+        'price': 200
+      },
+    ]
+  }, {
+    'type': 'Flight',
+    'offers': [
+      {
+        'id': 'luggage',
+        'title': 'Add luggage',
+        'price': 30
+      }, {
+        'id': 'comfort',
+        'title': 'Switch to comfort class',
+        'price': 100
+      }, {
+        'id': 'meal',
+        'title': 'Add meal',
+        'price': 15
+      }, {
+        'id': 'seats',
+        'title': 'Choose seats',
+        'price': 5
+      }, {
+        'id': 'train',
+        'title': 'Travel by train',
+        'price': 40
+      }
+    ]
+  }, {
+    'type': 'Restaurant',
+    'offers': [
+      {
+        'id': 'book',
+        'title': 'Book a table',
+        'price': 50
+      }
+    ]
+  }
+];
+
 
 // Функция из интернета по генерации случайного числа из диапазона
 // Источник - https://github.com/you-dont-need/You-Dont-Need-Lodash-Underscore#_random
@@ -66,15 +196,13 @@ const generatePhotos = () => {
 // RELATION BETWEEN POINT NAME - DESCRIPTION AND PHOTOS
 export const relationNameDescription = () => {
   const result = [];
-  let item = {};
-  for (const destination of DESTINATIONS) {
-    item ={};
-    item.destination = destination;
-    item.description = generateDescription();
-    item.photos = generatePhotos();
-    result.push(item);
-  }
-
+  DESTINATIONS.forEach((destination) => {
+    result.push({
+      destination: destination,
+      description: generateDescription(),
+      photos: generatePhotos(),
+    });
+  });
   return result;
 };
